+++ src/frontend/PIR.cpp
@@ -31,8 +31,6 @@ namespace panther{
 			}
 
 
-
-
 			return true;
 		};
 
@@ -40,7 +38,18 @@ namespace panther{
 
 
 		auto Type::operator==(const Type& rhs) const noexcept -> bool {
-			return this->base_type.id == rhs.base_type.id;
+			if(this->base_type.id != rhs.base_type.id){ return false; }
+
+			if(this->qualifiers.size() != rhs.qualifiers.size()){ return false; }
+
+			for(size_t i = 0; i < this->qualifiers.size(); i+=1){
+				const AST::Type::Qualifier& this_qualifiers = this->qualifiers[i];
+				const AST::Type::Qualifier& rhs_qualifiers = rhs.qualifiers[i];
+
+				if(this_qualifiers.is_ptr != rhs_qualifiers.is_ptr){ return false; }
+			}
+
+			return true;
 		};