+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### v0.14.0
+- Added pointer types
+- Added the `addr` keyword
+- Added the dereference (`.^`) operator 
+
 ### v0.13.1
 - Fixed `copy` expressions in globals
 - Fixed bug where compiler sometimes incorrectly errored that the entry function returned the incorrect type